Only look at 1 neighbourhood of neighbours
Reduce the amount of calculations performed when adding shortcuts by not
performing a 2 deep Dijkstra on the neighbourhood to determine if any
edges are shorter. The reasoning behind this is the assumption that a
shortcut added by this can be replaced later on by a better shortcut.
Thus this mechanism is also added to check on adding a shortcut whether
there already exits an edge between this nodes. If this is the case this
edge is simply replaced by the new shortcut.